Send internal SL to multiple clusters
Extracts the code added to the `send` command for this purpose into a
reusable function that is now also consumed by the `internal` command.
